requirements.txt
tensorflow>=2.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.23.0
flask>=1.1.2
Pillow>=8.0.0